Implemented the logic for discount coupons
Now the users can enter a valid coupon code and avail extra off on their cart items. The coupons can be created  in the admin page with a unique code, discount type, discount(percentage or amount), valid from, to and minimum purchase amount to use the coupon. When the coupon is used, the discount is applied on the existing final amount.